% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toWin.R
\name{toWin}
\alias{toWin}
\title{Convert windows file paths into R friendly ones}
\usage{
toWin()
}
\value{
The R friendly file path in the console and the clipboard
}
\description{
As the backslash has special properties in R (escape expression) one has to replace the backslashes in windows file paths with forwardslashes.
This is an annoying task and therefore there is this function. Just type toWin::toWin() into the console, paste the path into the console and get the R friendly path back (also in the Clipboard).
}
\author{
Emanuel Sommer
}
